+++ tests/SchemaContainerTest.php
@@ -14,7 +14,9 @@ namespace Hector\Schema\Tests;
 
 use Hector\Schema\Exception\NotFoundException;
 use Hector\Schema\Schema;
+use Hector\Schema\SchemaContainer;
 use Hector\Schema\Table;
+use Iterator;
 
 class SchemaContainerTest extends AbstractTestCase
 {
@@ -38,7 +40,7 @@ class SchemaContainerTest extends AbstractTestCase
         $schemaContainer = $this->getSchemaContainer();
         $iterator = $schemaContainer->getIterator();
 
-        $this->assertInstanceOf(\Iterator::class, $iterator);
+        $this->assertInstanceOf(Iterator::class, $iterator);
         $this->assertCount(1, $iterator);
         $this->assertContainsOnlyInstancesOf(Schema::class, $iterator);
     }
@@ -103,6 +105,18 @@ class SchemaContainerTest extends AbstractTestCase
         $this->assertInstanceOf(Table::class, $table);
         $this->assertEquals('customer', $table->getName());
         $this->assertEquals('sakila', $table->getSchemaName());
+    }
+
+    public function testGetTable_byAlias()
+    {
+        $schemaContainer = new SchemaContainer([
+            new Schema(connection: 'test', name: 'table1_name', charset: 'utf8mb4'),
+            $expected = new Schema(connection: 'test', name: 'table2_name', charset: 'utf8mb4', alias: 'myAlias'),
+            new Schema(connection: 'test', name: 'table3_name', charset: 'utf8mb4'),
+        ]);
+
+        $this->assertSame($expected, $schemaContainer->getSchema('myAlias'));
+        $this->assertSame($expected, $schemaContainer->getSchema('table2_name'));
     }
 
     public function testGetTableWithoutSchemaName()